requirements.txt
annotated-doc==0.0.4annotated-types==0.7.0anyio==4.12.0argon2-cffi==25.1.0argon2-cffi-bindings==25.1.0arrow==1.4.0asttokens==3.0.1async-lru==2.0.5attrs==25.4.0Babel==2.17.0beautifulsoup4==4.14.3black==25.11.0bleach==6.3.0certifi==2025.11.12cffi==2.0.0charset-normalizer==3.4.4click==8.3.1colorama==0.4.6comm==0.2.3contourpy==1.3.3cycler==0.12.1debugpy==1.8.17decorator==5.2.1defusedxml==0.7.1executing==2.2.1fastapi==0.123.5fastjsonschema==2.21.2filelock==3.20.0fonttools==4.61.0fqdn==1.5.1fsspec==2025.10.0h11==0.16.0httpcore==1.0.9httpx==0.28.1idna==3.11ipykernel==7.1.0ipython==9.7.0ipython-pygments-lexers==1.1.1ipywidgets==8.1.8isoduration==20.11.0jedi==0.19.2Jinja2==3.1.6joblib==1.5.2json5==0.12.1jsonpointer==3.0.0jsonschema==4.25.1jsonschema-specifications==2025.9.1jupyter==1.1.1jupyter-console==6.6.3jupyter-events==0.12.0jupyter-lsp==2.3.0jupyter-client==8.6.3jupyter-core==5.9.1jupyter-server==2.17.0jupyter-server-terminals==0.5.3jupyterlab==4.5.0jupyterlab-pygments==0.3.0jupyterlab-server==2.28.0jupyterlab-widgets==3.0.16kaggle==1.8.2kagglesdk==0.1.13kiwisolver==1.4.9lark==1.3.1librt==0.6.3MarkupSafe==3.0.3matplotlib==3.10.7matplotlib-inline==0.2.1mistune==3.1.4mpmath==1.3.0mypy==1.19.0mypy-extensions==1.1.0nbclient==0.10.2nbconvert==7.16.6nbformat==5.10.4nest-asyncio==1.6.0networkx==3.6notebook==7.5.0notebook-shim==0.2.4numpy==2.2.6opencv-python==4.12.0.88packaging==25.0pandas==2.3.3pandocfilters==1.5.1parso==0.8.5pathspec==0.12.1Pillow==12.0.0platformdirs==4.5.0polars==1.35.2polars-runtime-32==1.35.2prometheus_client==0.23.1prompt_toolkit==3.0.52protobuf==6.33.1psutil==7.1.3pure-eval==0.2.3pycparser==2.23pydantic==2.12.5pydantic-core==2.41.5Pygments==2.19.2pyparsing==3.2.5python-dateutil==2.9.0.post0python-json-logger==4.0.0python-multipart==0.0.20python-slugify==8.0.4pytokens==0.3.0pytz==2025.2pywinpty==3.0.2PyYAML==6.0.3pyzmq==27.1.0referencing==0.37.0requests==2.32.5rfc3339-validator==0.1.4rfc3986-validator==0.1.1rfc3987-syntax==1.1.0rpds-py==0.30.0schedule==1.2.2scikit-learn==1.7.2scipy==1.16.3seaborn==0.13.2Send2Trash==1.8.3setuptools==80.9.0six==1.17.0soupsieve==2.8stack-data==0.6.3starlette==0.50.0sympy==1.13.1terminado==0.18.1text-unidecode==1.3threadpoolctl==3.6.0tinycss2==1.4.0torch==2.6.0+cu124torchaudio==2.6.0+cu124torchvision==0.21.0+cu124tornado==6.5.2tqdm==4.67.1traitlets==5.14.3types-requests==2.32.4.20250913types-tqdm==4.67.0.20250809typing-inspection==0.4.2typing_extensions==4.15.0tzdata==2025.2ultralytics==8.3.234ultralytics-thop==2.0.18uri-template==1.3.0urllib3==2.5.0uvicorn==0.38.0wcwidth==0.2.14webcolors==25.10.0webencodings==0.5.1websocket-client==1.9.0widgetsnbextension==4.0.15